updated addPermitForm to work properly again (pulled from Tara's branch)
